+++ src/app.module.ts
@@ -19,7 +19,6 @@ import { DefaultStrategy } from "./modules/auth/basic-auth.strategy";
 import { TypeOrmModule } from "@nestjs/typeorm";
 import { getDatabaseConfiguration } from "./modules/database/db_configuration";
 import { UtilityController } from "./deprecated/utility.controller";
-import { AutomationService } from "./deprecated/automation.service";
 @Module({
   imports: [
     TypeOrmModule.forRoot(getDatabaseConfiguration()),
@@ -42,7 +41,6 @@ import { AutomationService } from "./deprecated/automation.service";
   providers: [
     { provide: APP_FILTER, useClass: LoggingExceptionFilter },
     DefaultStrategy,
-    AutomationService,
   ],
 })
 export class AppModule {}