+++ src/modules/files/files.controller.ts
@@ -1,0 +1,23 @@
+import { Controller, Put } from "@nestjs/common";
+import { ApiOperation, ApiOkResponse, ApiTags } from "@nestjs/swagger";
+import { Game } from "../games/game.entity";
+import { MinimumRole } from "../pagination/minimum-role.decorator";
+import { Role } from "../users/models/role.enum";
+import { FilesService } from "./files.service";
+
+@ApiTags("files")
+@Controller("files")
+export class FilesController {
+  constructor(private filesService: FilesService) {}
+
+  @Put("reindex")
+  @ApiOperation({
+    summary: "manually triggers an index of all games",
+    operationId: "indexGames",
+  })
+  @ApiOkResponse({ type: () => Game, isArray: true })
+  @MinimumRole(Role.ADMIN)
+  async reindex() {
+    return await this.filesService.index();
+  }
+}