+++ src/modules/files/files.module.ts
@@ -1,10 +1,13 @@
+import { FilesController } from "./files.controller";
 import { Module } from "@nestjs/common";
 import { FilesService } from "./files.service";
 import { GamesModule } from "../games/games.module";
+import { RawgModule } from "../providers/rawg/rawg.module";
+import { BoxartsModule } from "../boxarts/boxarts.module";
 
 @Module({
-  imports: [GamesModule],
-  controllers: [],
+  imports: [GamesModule, RawgModule, BoxartsModule],
+  controllers: [FilesController],
   providers: [FilesService],
   exports: [FilesService],
 })