+++ src/modules/games/games.controller.ts
@@ -1,5 +1,13 @@
-import { Controller, Get, Logger, Param, StreamableFile } from "@nestjs/common";
-import { ApiOkResponse, ApiOperation, ApiTags } from "@nestjs/swagger";
+import {
+  Body,
+  Controller,
+  Get,
+  Logger,
+  Param,
+  Put,
+  StreamableFile,
+} from "@nestjs/common";
+import { ApiBody, ApiOkResponse, ApiOperation, ApiTags } from "@nestjs/swagger";
 import { InjectRepository } from "@nestjs/typeorm";
 import {
   NO_PAGINATION,
@@ -19,6 +27,9 @@ import { FilesService } from "../files/files.service";
 import { GamesService } from "./games.service";
 import { MinimumRole } from "../pagination/minimum-role.decorator";
 import { Role } from "../users/models/role.enum";
+import { ImageUrlDto } from "../images/models/image-url.dto";
+import { ImagesService } from "../images/images.service";
+import { RawgIdDto } from "./models/rawg_id.dto";
 
 @ApiTags("game")
 @Controller("games")
@@ -28,6 +39,7 @@ export class GamesController {
   constructor(
     private gamesService: GamesService,
     private filesService: FilesService,
+    private imagesService: ImagesService,
     @InjectRepository(Game)
     private readonly gamesRepository: Repository<Game>,
   ) {}
@@ -149,5 +161,37 @@ export class GamesController {
   @ApiOkResponse({ type: () => StreamableFile })
   async downloadGame(@Param() params: IdDto): Promise<StreamableFile> {
     return await this.filesService.downloadGame(Number(params.id));
+  }
+
+  @Put(":id/box_image")
+  @ApiOperation({
+    summary: "updates the box art of a game",
+    operationId: "updateBoxArt",
+  })
+  @ApiOkResponse({ type: () => Game })
+  @ApiBody({ type: () => ImageUrlDto })
+  @MinimumRole(Role.EDITOR)
+  async remapBoxArt(
+    @Param() params: IdDto,
+    @Body() dto: ImageUrlDto,
+  ): Promise<Game> {
+    const game = await this.gamesService.getGameById(Number(params.id));
+    game.box_image = await this.imagesService.downloadImage(dto.image_url);
+    return await this.gamesService.saveGame(game);
+  }
+
+  @Put(":id/rawg_id")
+  @ApiOperation({
+    summary: "remaps a game to a rawg game and recaches it",
+    operationId: "remapGame",
+  })
+  @ApiOkResponse({ type: () => Game })
+  @ApiBody({ type: () => RawgIdDto })
+  @MinimumRole(Role.EDITOR)
+  async remapGame(
+    @Param() params: IdDto,
+    @Body() dto: RawgIdDto,
+  ): Promise<Game> {
+    return await this.gamesService.remapGame(Number(params.id), dto.rawg_id);
   }
 }